Add ELF symtab syms; our symtabs better store syms of same name; ELF/symtab refactoring.
Load more stuff from ELF info, including symtab.  Refactor all our prior
ELF support into a little library.  Change our symbol struct around
slightly to be able to support both DWARF and ELF symbols.  I think this
actually results in some of the structs being a little slimmed down.

How exactly we load in ELF symtab data is a bit weird.  Why?  Because
different Linux distros store ELF symtab in different areas.  All distros
must, of course, store dynamic symbol info in the actual binary/library,
because it's necessary for dynamic linking.  But, non-dynamic symtab data is
sometimes stripped out of binaries/libraries (technically it's not
necessary), and sometimes not.  If the non-dynamic symtab data is
stripped out, it should be in the debuginfo file.

So what we try to do is load the ELF symtab data first from the real
binary/library that is loaded into memory, and "associate" it into the
debugfile struct that also contains the DWARF info.  This is confusing,
but correct (that will be in an upcoming commit in target/).  For
internal use in dwdebug, we just try to load ELF symtab data from
whatever file the user has given us, IF the ELF symtabs haven't already
been filled in (the user could do this for us via the new ELF lib
functions, before loading the debugfile DWARF data).

Also, give our symtab struct the ability to cleanly store symbols with
the same name (which happens a fair bit).