Add support for optional warnings (vwarnopt).
vwarn (the non-optional warning macro) is still there.  Also, warnings
are conditioned in dumpdebuginfo via level (-w...w -- like -d...d) and
also on the set of log flags you might or might not pass in.  So the
debug and warning level numbers can be different, but for now, not the set
of flags.