Add symbol_get_location_offset and symbol_get_location_addr .
Quick methods to return error or statically-known location info
about a symbol.

Also, add a bit that says if, for a symbol, we were able to find
a base address during debuginfo parsing.  This is necessary for partially
loaded symbols, and we now have the bits to "waste" on it.