[LOCAL-DOS] Record the result of pass 1 in Makefile.
The purpose of pass 1 is to find out in what process the kernel panic
occured.  For that, it simply prints out task switch informations of
the replay session until the kernel crashes.  The result was following:

  ...
  4 (events/0) -> 285 (sh): brctr = 389344687
  285 (sh) -> 292 (sh): brctr = 389384499
  292 (local-kernel-pa) -> 285 (sh): brctr = 389460664
  285 (sh) -> 292 (local-kernel-pa): brctr = 389470539
  292 (local-kernel-pa) -> 293 (local-kernel-pa): brctr = 389521494
  293 (local-kernel-pa) -> 292 (local-kernel-pa): brctr = 411099251

Thus, the kernel crashed in the context of process 292 after the branch
counter 411099251.

We will use a "smarter" way to keep the results of analysis passes than
Makefile in the future, but not right now.