Make regions/addrs relocatable object<->real; refactor and use new ELF lib.
Regions and addresses are relocatable so we can handle dynamic linked
libs.  Regions get new fields: base_load_addr, base_phys_addr, and
base_virt_addr.  As we add ranges to the region, these fields are
updated.  Once a region has been fully loaded, we can compute the offset
between virtual (compiler) and physical (loaded) addresses.

Also, now we reuse a lot of the ELF interpretation code that had been
mixed into target and is now in dwdebug.

Finally, we now use the newer style of debuginfo loading that the
target API now requires.  We did this to handle the case where the
ELF symtab and DWARF debuginfo are actually in separate files.  Some
Linux distros put static ELF symtab symbols into the debuginfo file,
probably because of course that info is not needed for dynamic linking;
but others put static symbols into the actual binary along with
the dynamic symbols.  So we have to be a bit more agile when loading
ELF symtab symbols.

Also, properly check debugfile_load_opts that were associated with a
target by the user, and perhaps apply those options to specific
debuginfo files we might load (or, not load, due to the
debugfile_load_opts).