Add a tools subdir, and the first tool, psaction.
psaction allows you to do a bunch of things to processes running in
a VM.  Basically, you run it against a VM specified with the -m option,
then pass it a command (one of list|check|zombie|stop|kill), with an
optional argument.  list just lists all the processes.  check
takes an rfilter argument and returns any processes whose names are
accepted by the filter.  zombie|stop|kill all take an rfilter, and
act on processes whose names are rejected by the filter.  zombie sets
the `state' field in the PCB to -1 to make the process unrunnable (of
course, if it gets signaled, it can run again).  kill basically sets
up a pending SIGKILL for the process (and it assume a single-threaded
process for now; it also ignores various kernel locks (although these
ones I think were process-specific, so it's pretty safe)).  stop
tries to inject a signal via the same mechanism, but this doesn't
seem to work too well right now...