Add commands to enforce a process parentage predicate.
New commands: hiercheck and hierkill.  These commands load in the
process hierarchy, and check a command line-supplied list of regular
expressions that say which processes are valid.  In other words, any
process hierarchy strings (i.e., sh:getty:init) that match a regular
expression are "valid".  Otherwise, for a "hiercheck" command,
non-matching processes are printed out; for "hierkill", non-matching
processes are set up to be killed.