Add clrange_find_next_loosest().
Try to find the next index, given @index is (or might be) an existing
valid index in our range array -- BUT the index we find must *not* be
inside of the range(s) specified by @index; it must be completely
outside them.  Also, the bound we return will be the lengthiest
(loosest) bound; i.e., if there are multiple things at the next
index, we take the one with the widest range.

So, it's kind of the opposite of clrange_find(), which searches
inclusively for the tightest range containing the supplied @index.

This is useful so that, when searching a set of ranges that is populated
from the ELF symbol table, we can find the next "valid" symbol range.