Improvements to clrange_t: more kinds of searching; supports fast safe disasm.
Basically, clranges were designed to support fast address range lookups
for nested ranges (i.e., a function range, and multiple levels of nested
blocks within that range).

However, they were not designed to support an operation we need for
"safe" disasm, which is, given any address we want to start
disassembling at, find the previous known "safe" address (i.e., a
function symbol base address should always be safe).

Judy arrays, the structure that backs our clrange_t structure, don't
really have a concept of hierarchical ranges; we use them that way,
though, and the algorithm should be pretty fast that way.  So, since
they are flat, before these improvements we would have been unable to
knowingly find the "safest" previous symbol to start disassembling at.

What these functions give is is a way to ask, for any address X, what is
the "loosest containing" range (that is the `clrange_find_loosest'
function); or what is the *next* loosest containing range (i.e., the
loosest range that directly follows the loosest containing range of X).
This helps us locate, for any nesting depth of ranges, the "outermost"
containing range -- which in normal text segments, will be associated
with a function symbol or label, and is going to be safe to start
disassembly at.

If we had full DWARF descriptions for every last symbol and address in
the text segments of an executable, this would not be a problem, because
no range would be unknown -- i.e., every address would have a known safe
starting address for disassembly.  However, this is sadly not the case.

We could have relied on ELF symbols (ELF symbol coverage should be more
complete than DWARF symbols, but some ELF symbols don't really have size
info... so we have to hypothesize about them).  But at the same time, if
we don't have ELF symbol info, we'll have to use DWARF info, and that
means handling all the nested range info.  I suppose it might be
possible to have nested ELF symbols as well; this commit should cover
all that too.

The other thing this helps with is makes sure that when we load text
segment content from target memory to disassemble it, we are always
loading discrete chunks associated with symbols.  So, when somebody asks
for disasm at the gadget at addr <foofunc>+25, now we can just load all
of foofunc's code, and return the code at offset 25.  Also, if they ask
for disasm of a called address in the PLT when the called func is really
in a shared lib, now we can (or should soon be able to) then just load
the entire PLT and disassemble it all, rather than in chunks -- because
now we know where to start and where to end disasm.

Technical details: we now make sure each range -- when we add a range to
a clrange struct, or when we update an existing range's end -- has a
pointer to the range that directly contains it -- if there is one.
Without maintaining this pointer, such  a search through a Judy array
data structure seems to be impossible -- at least I couldn't come up
with an algorithm to do it!

Whewph.  But I think this is most of the icky gunk that underlies all
the "safe disassembly" stuff, which we need for CFI and the ROP
detection tools.