New routines supporting safe disasm and text segment loading/caching.
When we load text segment code to insert probes, we are usually going to
replace part(s) of the code with software breakpoints.  Thus, if some
library call tries later to disassemble the function containing the
inserted breakpoints by reading the text from the target again, their
disasm might fail, or will at least be "wrong".  Another problem is that
we don't want to simply load and cache arbitrary byte sequences from the
target when inserting breakpoints; this leads to a fragmented cache that
is hard to deal with.

So, we leverage our knowledge of the ELF symbol layout of the program to
only load and cache known "segments" of the code.  So, if the user asks
for loading a subset of a function we have not loaded into cache, we
actually load the whole thing, then return the subset.  The user can
request an independent copy of the code buf, OR, if possible, this
function will return a no-copy ptr into the cached code buffer (it is
possible when either the user requests an entire cached segment, OR if
they request an index into a cached segment with a length that is
segment_len - index).