Bug: linux_userproc_launch was not actually pausing the target after launch.
I must not have read the docs for PTRACE_DETACH fully, because
linux_userproc_launch was launching the target, but then letting it
resume before the rest of the target API functions could be called to
finish preparing the target.  _launch launches the user program and
traces it, looking for specific sequences of syscalls as the program is
loaded into memory and the address space is setup.  Once these sequences
have finished, then it returns.  The idea is that we have to wait for
the address space to get setup before we can initialize the target --
for instance, shared libs must be linked in.  BUT, _launch was
PTRACE_DETACH'ing -- which allows the traced process to resume.  So we
weren't actually waiting.

I added a target-generic flag, initdidattach, that a target creation
factory function (like linux_userproc_launch) can set if the create
function had to attach to the target, but had to leave it attached.  It
has to be a generic flag because the target-private data has not yet
been initialized.