Add the `come-from' ROP attack detector.
This tool and its support libs read in a file of gadgets and either
attach to a running target, or launch it, and then stick in gadget
probes.  When the gadget's RET is hit, we attempt to check that the
return address is a valid one.  Obviously, this only works for return
gadgets; this doesn't allow us to check JMP, SYSENTER, INT.  Also, for
now, it only *knows* that a gadget was used as part of proper control
flow if the CALL that invoked the code segment containing the gadget is
absolute (does not indirect through mem or register).  Not sure how bad
this little problem will bite us.

We support ROP gadgets that are unaligned to instruction boundaries in
the original program (i.e., that are embedded inside instructions).  So,
  1) gadget starts at real, and ends at real
  2) gadget starts in mid of real instr, then ends with real ret
  3) gadget starts in mid of real instr, then ends inside a real instr

When we install a gadget probe, we do several things:
  - determine which kind of gadget above we have:
    - find and disasm the safe disasm range containing the gadget
    - disasm the gadget alone
    - figure out if the gadget starts on an instruction boundary
  - if 2 or 3, create a probe on the first byte of the aligned
    instruction that precedes the gadget's first byte (the guard probe
    -- if this probe is hit, we have to immediately remove the probe on
    the first byte of the gadget).
  - otherwise, just create probes on the gadget's first byte, and on its
    RET instruction.

Then, whenever we arrive at the gadget entry probe, we set a flag
stating we're in the gadget.  When we hit the gadget's return, we grab
the return address off the top of the stack, and find both the text
segment containing the return address, and the text segment containing
our IP.  If the instruction preceding the return address was a CALL with
a static target address, and that address is *within the text segment
containing our gadget*, we assert that this is a PROPER use of the
gadget.  If the preceding instruction is not a CALL, we flag it as a
misuse of the gadget.

If the preceding instruction is a CALL, but it's an indirect call, for
now, it is an error -- we cannot tell (easily) if this was a proper use
of the gadget.  I'm not sure if there is a way to know for sure.