probe_hard_(enable|disable) replace probe_hard_(remove|insert) and now recurse.
The difference between hard disable/enable and regular enable/disable is that
enable/disable just makes sure the pre/post handlers are not run.  When a
probe (really, one of its underlying probepoints) is hard disabled, the
probepoint is actually removed from target.

Since probepoints can be shared among probes, probe_hard_disable takes a
`force' param; which caller can set this is vague :).  If force is not set,
the removal will fail.