Change location_resolve to be useful across the register/mem case.
Also, update various target address computation and load functions
to use the new thing.  Sometimes during address computation (well, when
the caller tries to auto deref a pointer or a string pointer), it's ok
if the value is technically in a register.  The value we are
*interested* in, in this case, is the pointed-to value.  If the value
really is in a register, we can't compute the address, of course.

So this should smooth things over more.