Fix bugs with hardware probepoints; improve the model for them too.
In the Xen target, there is a "mostly fake" "global" thread that is used
for two purposes: first, as the "real" thread when a debug exception has
to be handled when the kernel is in interrupt context, and therefore has
a stack, but is not associated with a real kernel thread (a task); and
second, for storing hardware probepoints for the TID_GLOBAL case -- i.e.,
hardware probepoints that should occur for all threads.

So... when we have both modifications to shared hardware probepoint
state, and to the current thread's other registers (like EIP), we must
carefully flush CPU context back to the VM, taking the non-debug registers
from the current thread (or the global thread if in interrupt context),
but then merging the debug registers for the current thread with those
from the global thread.  Also, whenever we are handling a global hardware
probepoint, we have to flush the status debug register for both the
global and current threads, to make sure this merge happens right when
we flush.

Now, right now, we only support global hardware probepoints; that check
is now there.  But, the overall concept of merging supports per-thread
hardware probepoints (and the Linux kernel supports them too, but we have
to go muck with thread state in an unpleasant way and rely on the kernel
to set the per-thread debug registers on context switch -- do it later).