Fix up ELF symbol size guessing.
This was broken... because I only fixed up the address range data struct!
I didn't actually change the size from 0 :).

But, it was strategically broken, too.  For asm symbols (which are the
0-length symbols we really care about), we were just looking for the
sequentially next (by address) ELF symbol.  For global ELF symbols (which
the Linux kernel (at least sometimes) uses to denote major function entry
points, if the major entry points had labels within them... we were assuming
that the global symbol ended at the label (a local ELF symbol).  This was
broken from the perspective that when we went to disasm those "function"
symbols, we would only disasm the first part, possibly.

So now we expand the heuristic: if we encounter a 0-length global ELF
function symbol (or one we assume is a function symbol because it's in a
text segment), we look for the next global symbol to mark as the real "end"
of the 0-length symbol.

This now lets system_call be "fully" disassembled, which was the original
goal.

Of course, now this is all useless to solve the original problem that
inspired it (catching syscall returns) for two reasons:

1) the irets that system_call "includes" can be jumped into from non-syscall
contexts, can return to a kernel thread, not a userspace thread...

2) we can't handle breakpoints on IRETs now because of the new multithread
support.  Once we single step the IRET, we are in a new thread!!!

Crap.  Now, the multithread stuff has to always check what it's single
stepping to see if it's stepping an IRET.  Fortunately, this is a cheap
single byte check.  Then what if the IRET faults while single stepping...

I must be crazy to spend my Saturday doing stuff like this.  Is anybody
even reading anymore?  :)