Add a hack to let users listen on function entry/return (RET|IRET).
If you pass <func_name>:E on the command line, you will get probes
placed on function entry (after prologue, if there is one), and on
function return (RET or IRET).  A pre-handler will be called on entry
that tries to dump args; a post-handler will be called on any of the
returns.