Support stepping IRETs, and add a time/step counter API.
First, we add support for detecting instructions that might change
context (i.e., change threads), and if/when we single step one of those
instructions, we save the old thread, assume the single step is going to
be executed atomically (not interrupted nor will it fault), and then when
we see the single step debug event, handle the event in the old thread --
not the new thread.  This makes sure that the old thread is notified before
the new thread gets control, which is probably the right thing in this case.

For the Xen target, this means being careful not to try to "load" any kernel
thread state if the EIP is a userspace VM address.

Also, we add a simple time API to support xentt.  Right now, we can return
the tsc, nanosecond time since boot, and the brctr.

Of course, tsc/nanosecond times from xen are read by mapping the shared info
page, not by actually reading the TSC.  But, Xen only updates these fields
periodically, so we might have to figure a more precise way to return these
values.  We don't want to read the tsc ourselves... maybe there is some other
way to get the current value of TSC so we can provide the correct, non-stale
values.

Also, fix up a bunch of debugging with better debug level values.