A bunch more at/until stuff.
Now, with '-A <func|addr>', we put a probe on the place, wait til it is hit,
then remove it and add the rest of our probes.  Also at that time, if the '-B'
option is passed, we enable the Xen BTS (if -B is passed without -A, we enable
BTS right away before doing any probing).

If '-U <func|addr>' is passed, we keep probing until that func|addr is hit;
if the '-B' flag was set, we stop BTS at this point; if the '-S' flag was set,
we also stop probing, remove all probes, cleanup, and exit.