Update base_addr attribute for symtabs (really for symtab symbols).
base_addr is a symbol attribute, but sometimes code ranges for a inlined
function instance (probably always) are specified in the next DWARF child
node, which is usually a lexical block or something that we represent with
a symtab.  So, we need to always be careful to propagate any range updates
from symtabs to update the base_addr of the symbol the symtab is associated
with.