Track syscall exits correctly.
This commit uncomments some stuff Chung had commented out until syscall exits
were tracked correctly.  To get this to work, I had to...

Add support for registering on IRETs, not just RETs (system_call returns
via IRET, of course).

When an IRET is hit, check to make sure 1) the coming-from context is not
interrupt context (obviously, if it is, we're returning somehwere (user or
kernel process context) after an interrupt handler, not a syscall; and
2) check the going-to context by reading the CS value off the stack that
the IRET is about to pop; if it's going to privilege level 3... then it must
be a syscall return.