Make sure to check for a hw breakpoint condition even if EIP is userspace.
This is insane.  By the time a kernel watchpoint actually causes a CPU
debug exception, a nontrivial amount of the time, the system has already
returned from the kernel, into userspace.  I knew watchpoints don't trigger
just at the time the read/write happens, but this is really bad.  Setting
the break-exact bits does not help (the manual basically says those only
work for older processors anyway) (they are suppose to "slow" teh CPU down
to give the writes a chance to happen, presumably closer to the actual
instruction that did the read/write).

So, we were checking to make sure we weren't trying to handle debug
exceptions in userspace (i.e., if EIP is a userspace address) because of
a different problem the other day.  Now, we have to be a bit more careful
and check if the breakpoint is a hardware one; if it is, just try to handle
it in the global thread, without loading any kernel threads.

NOTE: this will only work for global hardware watchpoints; when library
users really start using per-thread watchpoints, we're going to have to do
something else.

Anyway, this fixes watchpoints, I believe.  Yet more nontrivial CPU stuff...