Fix up CFI analysis so it works if inline instance functions are part of CF.
Either get_page wasn't inlined before, or something changed a bit with
how we handled inline instances after this was built.

Regardless, now this implementation of CFI stores symbols, not symbol
strings, for tracking its call stack.  We also try to track CFI based on
inline instance parents, not the inline instances themselves (the range
for an instance will not (probably) be the same as for the function it is
inlined in -- and what we're really tracking is calls/rets, which inlined
code doesn't do).

Also fixed one little (bad) bug about how we load EIP from a handler;
ALWAYS use the current TID from target_gettid(), not TID_GLOBAL, even
if you registered probes with TID_GLOBAL.