Handle single stepping into userspace, even if it was not expected!
Well, doh, to catch the kernel executing userspace code, we have to track
this.  So, again, if we expected a single step, even if we can't find EIP
in kernel space, handle it in the thread we were expecting to single step,
if there is one.

This is basically the same as the IRET case, except we weren't expecting
it.  I don't think this can go wrong, and I don't think there are any other
cases like this anymore...