Make gSOAP happy by requiring elements to be in strict sequential order.
Unfortunately, we're not starting from XSD, but rather from RNC.  When
I use trang to convert from RNC to XSD, before feeding the XSDs to gSOAP,
it seems to generate xs:choice restrictions even when no such choice is
called for (i.e., if you have a set of unordered elements in the
pattern)..  That screws up gSOAP and causes the C structs to be whacko.

gSOAP seems to believe that xs:choice restrictions mean that it should
generate C datatypes using tagged unions -- because you'll be encoding
one field or another.  That's a good thing, unless I misunderstand the
power of xs:choice.

I wonder if this is a place where RNC and XSD cannot map directly into
each other, or if this is some trang problem (unlikely)... or if I just
don't understand xs:choice -- that is most likely!

Anyway, I think this gives a complete set of dwdebug XML/gSOAP structs.

It *does* impose unrequired orderings on the XML subelements, but that's
ok because right now we're the only ones generating them.