Better handling for DWARF bit_size and byte_size attrs.
Realistically, most types could have either byte_size or bit_size,
and variables can too.  This improves the handling of these attrs;
but it does make the .size field of struct symbol bimodal -- so it
is very important to read it through one of the accessor functions.

Also, symbol_bytesize will round a bitsize symbol up to the nearest
byte.