Ok, really fix up symbol size handling this time.
This is hopefully the final solution to the previous commits, which
tried to store either bit_size or byte_size.  For bitfields, we really
need both, it seems.  So now we store bit_size and byte_size (when they
are both set, the byte_size is the one for the containing integral type).

So, now *all* the symbol size info (!) has been moved into the symbol
struct itself, and the size field has become a union of basically the
byte_size-only situation, and the bit_size/bit_offset situation.  This
allows the size union to handle byte_size-only variables and types;
and bit_size/bit_offset variables too (bitfields).

Of course, there are hefty restrictions on bit_size/bit_offset/containing
byte_size -- 10 bits, 10 bits, and 12 bits.  But that's all fine because
bitfields are only over integral types.

The last commit(s) also broke the target/ subdir; fixed.