Get rid of the stupid symbol and symtab refs, and add manual multiref mode.
This is the first way of doing it; because gSOAP doesn't grok XSD well at
all, it results in lots and lots and lots of mallocs during serialization.

I have a second strategy forthcoming, but I wanted to get this in the
record books -- and the second strategy does not really do the "right
thing" with its schema.