Second method of getting rid of symbol/symtab refs.
Now symbol and symtab elements can have an optional sref attr that
represents either a reference to another element in the xml doc,
or to a object that must still be fetched from the server.

This way seems likely to reduce memory wasteage due to better gsoap
struct representation, and it should also reduce the total number
of mallocs.

The drawback is that since XML 1.0 does not support attr choice,
it is schematically possible to have a symbol with all the "full"
symbol contents, but with its sref attr set rather than its sid attr.
This is ok; if we ever have to handle input symbols, we'll just
validate manually; we will never violate the schema's intent in our
output.