Add "concrete" thoughts on sharing debuginfo amongst multiple processes.
Since loaded debugfiles are memory-intensive in our implementation (to
facilitate fast lookups), it is definitely beneficial if we can share
them.

But sharing a massive debugfile struct with all kinds of dynamic,
interwoven structs is challenging.  First, we want a fast malloc
(ideally a slab allocator) because most of our data is structs of
known sizes (i.e., 64 byte symbols!).  Second, we need an allocator
that supports the traditional malloc interface, since we link with
libs (namely glib and judy) that require one.  Third, we need an
allocator that we can fix a base address for (so we can share that
addr via shm_open and mmap with other processes).

These are some quick thoughts on how to get there; I think this will
actually work without too much pain, assuming we can find an allocator
that satisfies the 3 constraints above.