Add support to gsoap to NOT merge in xs:include content.
Basically, when the XSD reader reads <xs:include ...> elements, it merges
the include'd content into the including XSD schema directly.  This means
I could not try to layer XSD files atop each other, and still fragment out
the gsoap modules on a per-XSD-file basis.  gsoap would end up creating
duplicate types in my hierarchy.  This behavior makes it hard to use gsoap's
notion of modules.

So, now there is a mapping file that basically sticks an #import line into
the wsdl2h-generated header file on any <xs:include> whose schemaLocation
attribute matches a mapping file entry -- and does not generate the
content for the <xs:include> (i.e., does not generate the type definitions,
but does allow them to be used, in the expectation that they will be defined
in a gsoap module (i.e., the one that was #import'd)).

Sigh.