Monitors provide an abstract, select()-based to threads and processes.
Suppose you have a multithreaded server that you want to be able to
spawn, control, and pass data to and from either threads and/or processes.
Well, this API lets you do that, via msg passing over pipes.  The other
benefit to it is that all waiting is select()-based.  That is how our event
loops are designed in the `target' library; thus, to unify waiting for
multiple things, our top-level wait-er (a monitor attached to a control
pthread) must be able to wait for anything, i.e. proxied RPC request I/O,
over a file descriptor.

In our servers, we have to spawn, monitor, control, and exchange msgs
with child threads and processes (threads might contain targets;
processes contain analyses or special targets).  We launch target monitor
threads in response to certain RPCs that create targets or analyses.

The target library model is one of blocking/nonblocking waits on FDs.  We
cannot change that, and we would be crazy to do so.  But then in order to
design servers that handle multiple targets/analyses, or even a single
target in a single thread, we have to convert a waitpid() to an FD via a
waitpipe... and we need to communicate with the child... we have to make
a top-level, select()-based monitoring entity.

Because of the desire to isolate targets/analyses from the server, and
the need to proxy the incoming RPC to the right handling thread (i.e.,
we have to pass an RPC destined for a ptrace target to the exact thread
that is monitoring that ptrace'd target) for execution there; BUT we then
have to pass it back to the server for encoding and transmission.

At the moment, this last limitation with the target handling thread
needing to pass the RPC response back to the request thread in the
server for transmission *could* be washed away -- because we can either
share the connection FD directly with a handling thread, or pass it
intact to a handling process.  HOWEVER, if we ever want to use SSL (i.e.,
make our servers real), we can share the SSL-ified connection FD directly
amongst threads, but we cannot pass it to another process.  Thus, we
have to build in all this bidirectional pipe stuff.

(Note that we still need unidirectional pipes to signal targets/analyses
that there is a pending request for them to service.)