Be more careful with synch/asynch ptrace control signals.
Now that we have asynchronous SIGCHLD monitoring via monitors and evloops,
it is worth it to sort of build defensive multithread bug detection into
the non-thread-safe target lib.  In some cases (i.e., when the user calls
target_resume or target_pause), those functions will do a blocking
waitpid() to wait for the control signal to hit the monitored thread.

But, if the lib is being used in a multithreaded app and some other thread
is running an evloop_run() for the evloop registered with the target, the
evloop's handler might get notified asynchronously instead of our blocking
waitpid().  This we would not want; so we add the ability to mark some control
signals as synchronous (the evloop handler will thus not attempt to handle
them).

This is still not thread safe, of course.