Add support for ActionEvents, and for Actions to be scheduled.
The way we do this right now is a bit unfortunate, but it's the model.
Right now, a client must implement the targetListener SOAP service,
and if they want to schedule an action when a probe is hit, they can
do it in the response they send back from the targetListener server
side.

It's really the only way with RPC to add an action (say singlestep)
when you're being notified that the probe's breakpoint has just been
hit (via the pre handler).  All we have is the response; the target
will resume and singlestep the original instruction and fire the post
handler before we could make a SchedAction call on the server (this
doesn't exist yet anyway, not that it is hard).

A targetListener server can affect the outcome of the probe or action
handlers via return code, as well.  The right kind of error code
can remove/cancel the probe or action.