Don't necessarily track all probes to be probe_free()d in target_free().
Now it's optional.  Higher-level probes that want to dealloc their own
probes must be free to do so.  target_free() was freeing all probes, and
of course this was wrong.

Eventually, we could separate target-tracked probes into probes that are
automatically unregistered, and automatically freed.  Probably the default
should be to always unregister any registered probe to make sure the target
is cleaned up.  Someday we'll do this, but for now this is good enough.