Split out ELF stuff into a binfile abstraction (to handle relocs).
Because of the way elfutils is, we have to carve out our own binfile
backend.  Plus, we must do all relocations and layout for relocatable
modules ourselves.  elfutils' libdwfl handles relocation only for
debuginfo section.  This means we have to relocate enough of the non-
debuginfo sections to grab the real symtab values (in case we don't
have debuginfo for all symbols).  Since users can search either the
binary's symtab OR search in debuginfo for addrs, or symbols (and
some symbols are only in one or the other), the symtab must be correct
too.

So, now debugfiles build atop binfiles, which wrap elfutils.