Relocate kernel modules -- both the ELF stuff, and the debuginfo.
elfutils doesn't relocate all the ELF section data; it only relocates
debuginfo sections.  So, since some symbols we (might) need come from the
symtab, we have to relocate those manually.

For now, we do relocations in a memory image of the ELF file on
disk, and convince elfutils to use it.  This is not totally clean, but
it works fine for now.  At least we're not scattering temp files about.