Allow "bad" singlesteps to be inferred.
If the Xen kernel we're running on does not set the right bit in the
debug status register signaling a singlestep has occurred, and if
we were expecting a singlestep, and if the TF flag is set, and if the
IP is <= 15 bytes away from the probepoint we were expecting to singlestep
(max instr len), allow it to be processed as a valid singlestep.

We could verify by disassembling, but that is too much effort for now.