Broad improvements, mostly involving refcnts.
Improved refcnt debugging, a lot.  We're now more consistent with how
the libraries use refcnts internally, vs how users interact with
refcnt'd objects as library users.  For any function that a user calls
that holds a ref to an object, they must call *_release() on that
object.  The libraries themselves must *not* call _release() unless they
have a really good reason.  See the comments in include/common.h for
much more detail.

I added some of the weak ref stuff (eventually we'll use these to handle
circular refs, but for now it's not really necessary); it's not really
implemented yet though.

I moved some header stuff around and renamed some functions to better
delineate which ones hold refs to their return values (on behalf of a
user), and usually I added library-internal versions so that the library
can hold the objects appropriately, vs self-holding them (the user gets
self-held objects so they don't have to know about the details of
refcnting).

I fixed a whole bunch of memory bugs as I redid the RHOLD/RPUT macros.
The biggest one is in the Xen target; we were double-loading some thread
machine state in the internal debug interrupt handler, unintentionally,
and we were overwriting target values without freeing them.  Took a long
time to figure this out; thread loading inside the Xen target is too
complicated.  But now it's squeaky clean.

Oh, I also got rid of the unsightly DWDEBUG_USE_STRTAB.  It has been
replaced (in the negation/inversion sense) by DWDEBUG_NOUSE_STRTAB).  We
basically always want to cache the strtabs in ELF files we load; there
is no reason to malloc for each string in the table.  But I also made it
so that it's irrelevant to users, and only needs to be enabled when
building the dwdebug/ subdir.  Hopefully this also speeds up the
deallocation time for a debugfile; there are fewer checks to make before
choosing to free a string or not.