Fix relocation.
Well, this proves that my manual relocation was working, I think.  Turns
out I goofed the check for which sections *not* to relocate (remember,
elfutils relocates *only* the debug sections); so I was relocating all
sections that had a relocation section.  So, I was double-relocating the
debug sections ;).  Ooops.  Also, then, undo the "hack" that was trying
to understand elfutils to figure out what the heck was going on.  Also
some debugging.